CDAP-2169: Delete the app if the last adapter got deleted
